Added a lot of things; be warned [SEE BELOW FOR DETAILS]
Not sure if everything's working. I did end up implementing the library that I made as my personal project for the portfolio; oddly, I did so only to use a timer, which is not at all what my library is for, but is useful for my case anyway. Feel free to comment out the Timer object as well as any implementations, i.e. the entire contents of the draw loop and some portions of setup().
Everything that was already there before should still work. I did some rearranging of some of the code because it was getting kind of crowded. I also added a feature where it takes a random sample of 50 of the cleaned images and dispays them.
The machine learning/AI part, however, still needs some work. The main thing that I'm missing is a cost function, something that would tell the computer that it still needs to perfect its algorithm further.