Everything should work now
Still using the FunGUI library to help time images that are displayed from a random sample of images. Using the technique of clumping data that I devised has no benefit to the system so I'm going to find a different way to do that.
The main weak point here is the lack of a cost function, as I mentioned before. We can parse the data, but we need to be able to define some parameters that are really going to tell us whether the characters fit or not.